first rough, but working version of the serial lcd handling along with ethernet stuff
we're talking about <=128byte packets, for which i think it is
reasonable to send as big packets through the wire as possible because
the protocol overhead is comparable to the data size. so, from now on,
the answer coming from the uC is now sent in a single tcp packet.

the introduced apis use behave differently than Interaction regarding
instantiation and stuff. this is partly due to the fucked up nature of
how arduino handles classes, but also because it is more straightforward
this way (as a programmer). i need to re-think how these should work,
but right now the main priority is getting stuff done and masturbating
to code quality is absolutely secondary (although important).